Refactor: Move eBPF inspection commands to top-level
Move eBPF map inspection commands from 'system-probe usm ebpf' to
'system-probe ebpf' to make them available as cross-cutting utilities
for debugging all eBPF-based modules (USM, network tracer, OOM kill
probe, etc.).

Changes:
- Create new cmd/system-probe/subcommands/ebpf/ package
- Move map list/dump commands to top level
- Update documentation to reflect new command structure
- Remove ebpf subcommand from USM commands

Commands:
- system-probe ebpf map list
- system-probe ebpf map dump {id|name} <value>
