+++ cmd/system-probe/subcommands/ebpf/command_linux.go
@@ -5,7 +5,7 @@
 
 //go:build linux
 
-package usm
+package ebpf
 
 import (
 	"encoding/json"
