+++ cmd/system-probe/subcommands/ebpf/command_linux_test.go
@@ -5,7 +5,7 @@
 
 //go:build linux_bpf
 
-package usm
+package ebpf
 
 import (
 	"bytes"
