+++ cmd/system-probe/subcommands/subcommands.go
@@ -11,6 +11,7 @@ import (
 	cmdconfig "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/config"
 	cmdcoverage "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/coverage"
 	cmddebug "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/debug"
+	cmdebpf "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/ebpf"
 	cmdmodrestart "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/modrestart"
 	cmdrun "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/run"
 	cmdruntime "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/runtime"
@@ -29,6 +30,7 @@ func SysprobeSubcommands() []command.SubcommandFactory {
 		cmdconfig.Commands,
 		cmdruntime.Commands,
 		cmdcoverage.Commands,
+		cmdebpf.Commands,
 		cmdusm.Commands,
 	}
 }
